Move role-change notification inside members list
